requirements.txt
adafruit-io==2.7.2
flake8==7.1.0
python-dotenv==1.0.1
